pyproject.toml
[project]
name = "git-code-ai-analyzer"
version = "0.1.0"
description = "AI-assisted Git commit and code change analyzer"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "AaronXu", email = "mochi72@gmail.com"}
]
dependencies = [
    "gitpython",
    "requests",
    "openai"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"